refactor: move auto code into `Dataset`
