+++ hub/__init__.py
@@ -1,5 +1,4 @@
 from .api.dataset import Dataset
-from .auto.ingest import from_kaggle, from_path
-from .auto.load import load
+from .util.load import load
 
 __version__ = "2.0a0"
