+++ hub/api/tests/dummy_data/cat.jpeg
