+++ hub/api/tests/dummy_data/image_classification/class0/sample0.jpeg
