+++ hub/api/tests/dummy_data/image_classification/class1/sample0.jpeg
