+++ hub/api/tests/dummy_data/image_classification/class2/sample0.jpeg
