+++ hub/api/tests/dummy_data/image_classification_with_sets/test/class0/sample0.jpeg
