+++ hub/api/tests/dummy_data/image_classification_with_sets/train/class1/sample0.jpeg
